rename layout component to Layout
