+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
-import Layout from '../components/Layout'
+import Layout from '../components/Layout/Layout'
 
 
 function MyApp({ Component, pageProps }: AppProps) {
@@ -8,6 +8,7 @@ function MyApp({ Component, pageProps }: AppProps) {
     <Layout>
       <Component {...pageProps} />
     </Layout>
-)}
+  )
+}
 
 export default MyApp
